requirements.txt
asgiref==3.3.4
Django==3.2
django-bootstrap3==15.0.0
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2-binary==2.8.6
pytz==2021.1
simplejson==3.17.2
sqlparse==0.4.1
